restructure directory
