+++ test/query/index.test.js
@@ -1,0 +1,5 @@
+import { gqlFetchDetail, gqlFetchList } from "../../src/common";
+
+describe("Qewl#query utility functions", () => {
+  it("Qewl#gqlFetchDetail returns expected query", () => {});
+});
